64ビット用にコンパイルする際の警告を修正．

+++ mist/io/raw.h
@@ -277,7 +277,7 @@ namespace __raw_controller__
 				if( progress != pre_progress )
 				{
 					pre_progress = progress;
-					if( !f( progress ) )
+					if( !f( static_cast< double >( progress ) ) )
 					{
 						fclose( fp );
 						return( false );
